Initial Structure Complete
Wrote the initial code for the countdown() callback function.

Wrote the base functionality for starting and stopping the timer.

Still need to add the appropriate classes for use in the JS code, but these have been earmarked inside styles.css.

Attached jQuery to the project and added the reference in pomodoro.html.