New logic and structure!
I added loads of new logic to background.js, including functions to
access and alter the variables from the timer-script.js. I got the
backgroundChecker function up and running. The timer, break length, and
session length numbers are fetched from background.js instead of being
rendered on the page.

I also began cleanup on the html file, backing it up just in case. It
still has a lot of unnecessary data attributes and could stand to be
cleaned further.