I added notifications!
Now, whenever the swapState function is called, it adds a system
notification that informs the user about the current state of the timer
without requiring the user to open the popup.

Unfortunately, this makes a sound that happens at the same time as my
chime sound. I can't currently remove this, as the silent: true option
is currently under construction. I may have to remove this feature,
though it really depends on the user.

I also deleted files that are no longer being used, so I can prepare for
packaging/publication.